moderation-inference/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
onnxruntime==1.18.1
pillow==10.4.0
numpy==1.26.4
requests==2.32.3
python-multipart
